+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## 0.6.2 - Unreleased
 
+### Added
+
+- `Zoi.literal/2` type to accept only a specific literal value
+
 ### Changed
 
 - Refactor all errors to be generated on type creation instead of parsing time